fix(admin): extract producer types to db-free schema module
Client was pulling src/domains/admin/producers.ts (which imports db → pg →
fs) into the browser bundle. Moved constants + types to a sibling
producers-schema.ts; the server-only query file re-exports them so
server callers keep their single-file import path. Added 'server-only'
to the main file so any future regression fails at build time.

Fixes Turbopack 'Module not found: fs' build failure on CI.
