+++ src/domains/admin/producers.ts
@@ -1,6 +1,31 @@
+import 'server-only'
 import { db } from '@/lib/db'
 import type { VendorStatus } from '@/generated/prisma/enums'
 import { requireAdmin } from '@/lib/auth-guard'
+import {
+  DEFAULT_PAGE_SIZE,
+  PRODUCER_SORT_KEYS,
+  PRODUCER_STATUS_FILTERS,
+  type EnrichedProducer,
+  type ProducerSortKey,
+  type ProducerStatusFilter,
+  type ProducersOverview,
+  type ProducersOverviewParams,
+} from './producers-schema'
+
+// Re-export the schema surface so existing callers (page.tsx, tests) keep
+// their single-file import path while the DB-free types live in a sibling
+// module the client can safely import.
+export {
+  DEFAULT_PAGE_SIZE,
+  PRODUCER_SORT_KEYS,
+  PRODUCER_STATUS_FILTERS,
+  type EnrichedProducer,
+  type ProducerSortKey,
+  type ProducerStatusFilter,
+  type ProducersOverview,
+  type ProducersOverviewParams,
+}
 
 // Order statuses that count as "billed revenue" for a producer.
 // REFUNDED/CANCELLED are intentionally excluded.
@@ -13,86 +38,6 @@ const BILLED_STATUSES = [
 ] as const
 
 const SPARKLINE_DAYS = 14
-
-export const DEFAULT_PAGE_SIZE = 20
-
-export const PRODUCER_STATUS_FILTERS = [
-  'ALL',
-  'ACTIVE',
-  'APPLYING',
-  'PENDING_DOCS',
-  'SUSPENDED_TEMP',
-  'SUSPENDED_PERM',
-  'REJECTED',
-] as const
-export type ProducerStatusFilter = (typeof PRODUCER_STATUS_FILTERS)[number]
-
-export const PRODUCER_SORT_KEYS = [
-  'revenueDesc',
-  'revenueAsc',
-  'recent',
-  'lastSeen',
-  'name',
-  'orders',
-] as const
-export type ProducerSortKey = (typeof PRODUCER_SORT_KEYS)[number]
-
-export interface ProducerSparkPoint {
-  day: string
-  revenue: number
-}
-
-export interface EnrichedProducer {
-  id: string
-  slug: string
-  displayName: string
-  email: string
-  status: VendorStatus
-  description: string | null
-  location: string | null
-  logo: string | null
-  productsCount: number
-  stripeOnboarded: boolean
-  avgRating: number | null
-  totalReviews: number
-  createdAt: string
-  revenue: number
-  ordersCount: number
-  topProduct: { id: string; name: string; unitsSold: number } | null
-  lastSeenAt: string | null
-  sparkline: number[]
-}
-
-export interface ProducersOverviewParams {
-  page?: number
-  search?: string
-  status?: ProducerStatusFilter
-  sort?: ProducerSortKey
-}
-
-export interface ProducersOverview {
-  pageItems: EnrichedProducer[]
-  pagination: {
-    page: number
-    pageSize: number
-    totalFiltered: number
-    totalPages: number
-  }
-  params: {
-    search: string
-    status: ProducerStatusFilter
-    sort: ProducerSortKey
-  }
-  globals: {
-    total: number
-    active: number
-    pendingReview: number
-    suspended: number
-    gmv: number
-    orders: number
-  }
-  statusCounts: Record<VendorStatus, number>
-}
 
 interface RevenueRow {
   vendorId: string
